requirements.txt
requests
beautifulsoup4
pandas
matplotlib
alpha_vantage
tkinter
seaborn
pillow
scikit-learn